Update IndexStatus methods
