+++ src/test/scala/io/qbeast/spark/index/NormalizedWeightIntegrationTest.scala
@@ -75,8 +75,7 @@ class NormalizedWeightIntegrationTest extends QbeastIntegrationTestSpec {
           .save(tmpDir)
 
         val qbeastSnapshot = getQbeastSnapshot(tmpDir)
-        val cubeNormalizedWeights =
-          qbeastSnapshot.loadLatestIndexStatus.cubeNormalizedWeights
+        val cubeNormalizedWeights = qbeastSnapshot.loadLatestIndexStatus.cubeNormalizedWeights()
 
         cubeNormalizedWeights.foreach(cubeInfo => cubeInfo._2 shouldBe 2.0)
     }
